refactor(profiles): centralize animation duration sync + clamp
Review follow-ups for #51.

- Add a single settings->engine boundary (applyTilingSettings,
  setAnimationDuration, setScrollSpeed in KiwiCore+AnimationSync)
  so the engine's cached durations can't diverge from the
  persisted TilingSettings. Route the three whole-settings sites
  (apply(profile:), apply(composed:), applyProfileScopedState)
  and the four command dual-writes through it. Fixes the GUI
  live-apply path retiling at a stale duration.
- Clamp durationMS/scrollSpeedMS to 50-1000 at the
  AnimationSettings boundary (didSet + decode) so a persisted or
  hand-edited value never diverges from the engine.
- Split KiwiCore+MonitorChange out of KiwiCore+ProfileResolution
  (was at the 350-line ceiling).
- Extend AnimationPersistenceTests: settings-side clamp, decode
  clamp, and the GUI-apply engine sync.
