+++ Sources/KiwiDeskCore/Tiling/AnimationSettings.swift
@@ -42,16 +42,29 @@ public struct AnimationSettings: Sendable, Equatable, Codable {
     /// whole layout. Named to avoid the `layout_change` event.
     public var onRelayout = true
 
-    /// General spring animation duration (ms), clamped 50–1000
-    /// by `AnimationEngine`. Synced to the engine on profile
-    /// apply. Default 250. Lua: `animations.set_duration`.
-    public var durationMS = 250
+    /// General spring animation duration (ms). Clamped 50–1000
+    /// here (the single clamp authority) so a persisted or
+    /// hand-edited value can never diverge from what the engine
+    /// runs. Synced to the engine on profile apply. Default 250.
+    /// Lua: `animations.set_duration`.
+    public var durationMS = 250 {
+        didSet { durationMS = Self.clampMS(durationMS) }
+    }
 
     /// Scrolling-layout focus-shift animation speed (ms),
     /// clamped 50–1000. Separate from `durationMS` so each
     /// knob tunes its own trigger. Synced on profile apply.
     /// Lua: `animations.set_scroll_speed` (issue #51).
-    public var scrollSpeedMS = 250
+    public var scrollSpeedMS = 250 {
+        didSet { scrollSpeedMS = Self.clampMS(scrollSpeedMS) }
+    }
+
+    /// The supported animation-duration range (ms). Kept in
+    /// lockstep with `AnimationEngine`'s clamp; a value written
+    /// here and mirrored onto the engine lands identically.
+    static func clampMS(_ ms: Int) -> Int {
+        min(max(ms, 50), 1000)
+    }
 
     private enum CodingKeys: String, CodingKey {
         case onSpaceChange = "on_space_change"
@@ -97,15 +110,19 @@ public struct AnimationSettings: Sendable, Equatable, Codable {
                 Bool.self,
                 forKey: .onRelayout
             ) ?? true
-        durationMS =
+        // `didSet` observers don't fire during init, so clamp
+        // the decoded values explicitly through the same helper.
+        durationMS = Self.clampMS(
             try container.decodeIfPresent(
                 Int.self,
                 forKey: .durationMS
             ) ?? 250
-        scrollSpeedMS =
+        )
+        scrollSpeedMS = Self.clampMS(
             try container.decodeIfPresent(
                 Int.self,
                 forKey: .scrollSpeedMS
             ) ?? 250
+        )
     }
 }
